+++ apps/web/src/pages/Auth/RegistrationPages/GettingStartedPage/messages.ts
@@ -16,6 +16,11 @@ const messages = defineMessages({
     id: "g+tCWe",
     description: "Subtitle for the registration pages",
   },
-});
+  gettingStartedSectionTitle: {
+    defaultMessage: "Getting started",
+    id: "QXiUo/",
+    description: "Main heading in getting started page.",
+  },
+} as const);
 
 export default messages;
