+++ apps/web/src/pages/Auth/RegistrationPages/RegistrationContactEmailVerificationPage.tsx
@@ -11,7 +11,7 @@ import useRoutes from "~/hooks/useRoutes";
 import RequireAuth from "~/components/RequireAuth/RequireAuth";
 import EmailVerificationApi from "~/components/EmailVerification/EmailVerification";
 
-import messages from "./utils/messages";
+import messages from "./GettingStartedPage/messages";
 
 const RegistrationContactEmailVerificationPage = () => {
   const intl = useIntl();
